added: error boundary 적용
