+++ todolist-frontend/src/components/ErrorBoundary/index.js
@@ -1,34 +1,1 @@
-import React, { Component } from 'react';
-
-class ErrorBoundary extends Component {
-  state = {
-    error: null,
-    errorInfo: null
-  };
-
-  componentDidCatch(error, errorInfo) {
-    this.setState({
-      error: error,
-      errorInfo: errorInfo
-    });
-  }
-
-  render() {
-    if (this.state.errorInfo) {
-      return (
-        <div>
-          <h2>Something went wrong.</h2>
-          <details style={{ whiteSpace: 'pre-wrap' }}>
-            {this.state.error && this.state.error.toString()}
-            <br />
-            {this.state.errorInfo.componentStack}
-          </details>
-        </div>
-      );
-    }
-
-    return this.props.children;
-  }
-}
-
-export default ErrorBoundary;
+export { default } from './ErrorBoundary';
